fix(ansible-docker): Fix mtu value of custom docker networks (#105)
The current bridge networks on the deployed containers do not have their mtu value set to match the openstack host value. It was mistakenly thought that setting the value in the /etc/docker/daemon.json would cause all new networks to inherit this value but it actually only affects the default bridge network.