feat(infra): Create a docker-based local development stack (#107)
### Summary

Creates a stack to emulate the deployed stack using docker compose. This
is mostly intended for rapid development of the pipelines without having
to test them using remote services.

Note that there is a reasonable level of duplicated configuration code
between the local & deployed services so that the code is not overly
complex.

Fixes #98