fix(pipelines/ingest) - Allow dlt.resource or dlt.source decorated functions in cli_main (#112)
### Summary

A pipeline object is required in order for the dlt configuration
framework to use the pipeline name to further disambiguate keys. Without
a pipeline the config framework stops at the left most `sources` key.

In `cli_main` we check if the `data_generator` needs calling before
being passed to `pipeline.run`.