feat(warehouses/accelerator): Mean time between failures model (#139)
### Summary

Implements a dbt model to compute the [mean time between failures
metric](https://en.wikipedia.org/wiki/Mean_time_between_failures).

Much of this work included a restructuring of the staging models to
better hide all of the Opralog relational tables.

There was also a fix to the accelerator sharepoint timestamps. They were
thought to be UTC but are actually local so need to be converted to UTC.