fix(warehouses/accelerator): Set first uptime value to NULL for first dowtime record of cycle (#143)
### Summary

It's not correct to arbitrarily take the cycle start as the first
equipment up point as some equipment may be on all of the time. Rename
the `uptime_mins` to `uptime_before_fault_mins` for clarity and set it
to NULL for the first entry in a cycle.

For the MTBF calculation equipment with fewer than 3 faults will be
ignored in Superset.